% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataOrganization.R
\name{Code}
\alias{Code}
\title{Code Parsed Data}
\usage{
Code(parsedData)
}
\arguments{
\item{parsedData}{A list containing the values, key, and all possibilities. Created by the ParseData function.}
}
\description{
Coded Data parsed by the ParseData function in the TEI package.
}
\examples{
### Pasrse the data
parsedData <- ParseData(resps,key)

### Code the data based on the keys
codedData <- Code(parsedData)
}
\keyword{code}
