% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataOrganization.R
\name{CorSmooth}
\alias{CorSmooth}
\title{Smoothed Correlation Matrix}
\usage{
CorSmooth(x, ridge = 1e-04)
}
\arguments{
\item{x}{A matrix requiring smoothing}
}
\description{
Smooth a non positive definite matrix by a brute force algorithm.
}
\examples{

mat <- CorSmooth(x)
}
