% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoringAlgorithms.R
\name{MCC}
\alias{MCC}
\title{Matthews correlation coefficient}
\usage{
MCC(keyNum, correctCheckNum, incorrectCheckNum, distractorNum)
}
\arguments{
\item{keyNum}{The number of correct response options in the key.}

\item{correctCheckNum}{The number of correct response options marked by the examinee}

\item{distractorNum}{The number of innocrrect response options in the key.}

\item{incrrectCheckNum}{The number of incorrect response options marked by the examinee}
}
\description{
Calculates the Matthews correlation coefficient as a scoring algorithm.
}
\examples{
MCC(keyNum = 4, correctCheckNum = 4, incorrectCheckNum = 0,distractorNum = 4)
}
\keyword{scoring}
