% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoringAlgorithms.R
\name{Morgan}
\alias{Morgan}
\title{The Morgan scoring algorithm}
\usage{
Morgan(keyNum, correctCheckNum, incorrectCheckNum, distractorNum)
}
\arguments{
\item{keyNum}{The number of correct response options in the key.}

\item{correctCheckNum}{The number of correct response options marked by the examinee}

\item{distractorNum}{The number of innocrrect response options in the key.}

\item{incrrectCheckNum}{The number of incorrect response options marked by the examinee}
}
\description{
Implements the scoring algorithm described in Morgan (1979). Penalizes incorrect responses and rewards correctly marked responses. Also known as the 'Middlesex Scoring Scheme', discussed in detail in Buckly-Sharp and Harris (1971).
}
\examples{
Morgan(keyNum = 4, correctCheckNum = 4, incorrectCheckNum = 0,distractorNum = 4)
}
\keyword{scoring}
