% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataOrganization.R
\name{Nperm}
\alias{Nperm}
\title{Number of permutations}
\usage{
Nperm(n, r)
}
\arguments{
\item{n}{The total number of choices}

\item{r}{The number of correct response options marked by the examinee}
}
\description{
Calculates the number of permutatons of a given length for a larger collection of objects.
}
\examples{
Nperm(10,6)
}
\keyword{permutation}
