% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataOrganization.R
\name{NumberPermsAnySize}
\alias{NumberPermsAnySize}
\title{Total Number of Permutations of all sizes}
\usage{
NumberPermsAnySize(n, min = 1)
}
\arguments{
\item{n}{The total number of choices}

\item{min}{The total number of choices}
}
\description{
Calculates the number of permutatons for a total number of objects when it is possible to select any number of them (with no repeats). This is useful for calculateing how many scoring possibilities for a  click and drag ordering item. Also allows for specifying a minimum number. Such as for items of the form: Select at least 5 most important things to do, and order them by priority.
}
\examples{
NumberPermsAnySize(10,6)
}
\keyword{permutation}
