% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataOrganization.R
\name{OnlyLettersNumbers}
\alias{OnlyLettersNumbers}
\title{Default Parsers}
\usage{
OnlyLettersNumbers(x)
}
\arguments{
\item{x}{a string to parse}
}
\description{
An example (default) parser for the ParseData function.
}
\examples{
### Pasrse the data
parsedData <- ParseData(resps,key)

### Code the data based on the keys
codedData <- Code(parsedData)
}
\keyword{code}
