% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoringAlgorithms.R
\name{plusMinus}
\alias{plusMinus}
\title{Plus/Minus Scoring Algorithm}
\usage{
plusMinus(keyNum, correctCheckNum, incorrectCheckNum, distractorNum)
}
\arguments{
\item{keyNum}{The number of correct response options in the key.}

\item{correctCheckNum}{The number of correct response options marked by the examinee}

\item{distractorNum}{The number of innocrrect response options in the key.}

\item{incrrectCheckNum}{The number of incorrect response options marked by the examinee}
}
\description{
SA 2 from Domnich et al (2015). Recieves the addition of
}
\examples{
plusMinus(keyNum = 4, correctCheckNum = 4, incorrectCheckNum = 0,distractorNum = 4)
}
\keyword{scoring}
