% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoringAlgorithms.R
\name{Ripkey}
\alias{Ripkey}
\title{Ripkey Partial Credit Scoring}
\usage{
Ripkey(keyNum, correctCheckNum, incorrectCheckNum, distractorNum)
}
\arguments{
\item{keyNum}{The number of correct response options in the key.}

\item{correctCheckNum}{The number of correct response options marked by the examinee}

\item{distractorNum}{The number of innocrrect response options in the key.}

\item{incrrectCheckNum}{The number of incorrect response options marked by the examinee}
}
\description{
Partial credit scoring algorithm described in Ripkey et al (1996) for TEIs.
Evaluates to the number of correctly marked response options divided by the total number of keyed-correct options. If more than the keyed number of options are marked, the resulting score is zero.
}
\examples{
Ripkey(keyNum = 4, correctCheckNum = 4, incorrectCheckNum = 0,distractorNum = 4)
}
\keyword{scoring}
