% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataOrganization.R
\name{subCor}
\alias{subCor}
\title{Subset a correlation matrix}
\usage{
subCor(mat, x)
}
\arguments{
\item{mat}{The square matrix to subset}

\item{x}{The required (or unrequired) column indexes.}
}
\description{
Create a subset of a square matrix by column indexes. Allows for either plus or minus.
}
\examples{
### Select a few columns
subCor(mat,c(4,6,8))
### Removed a few columns
subCor(mat,-3)
}
\keyword{subset}
