% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoringAlgorithms.R
\name{TrueFalse}
\alias{TrueFalse}
\title{True-False Scoring Algorithm}
\usage{
TrueFalse(keyNum, correctCheckNum, incorrectCheckNum, distractorNum)
}
\arguments{
\item{keyNum}{The number of correct response options in the key.}

\item{correctCheckNum}{The number of correct response options marked by the examinee}

\item{distractorNum}{The number of innocrrect response options in the key.}

\item{incrrectCheckNum}{The number of incorrect response options marked by the examinee}
}
\description{
Formula 4 (Husa et al, 1984) or SA 3 (Domnich et al, 2015). Simply the total number of options appropriately marked divided by the number of options. Therefore, has no penalty marking an incorrect item. Equivalent to the proportion of correct responses if each choices was scored individually as a True-False item. Hence, the name.
}
\examples{
TrueFalse(keyNum = 4, correctCheckNum = 4, incorrectCheckNum = 0,distractorNum = 4)
}
\keyword{scoring}
